Conflicts:
	gmxtest.pl

Changed ethyleenglycol dt to 0.001 from 0.002 because the test seemed
to unstable. ICC gave different results with and without -mavx.
New reference values generated with Gromacs 4.5.5, gcc 4.7.1.

+++ gmxtest.pl
@@ -311,7 +311,7 @@ sub test_systems {
                         $nerror++;
                         next;
                       }
-		      $line1 =~ s/(e[-+])0([0-9][0-9])/\1\2/g; #hack on windows X.Xe-00X -> X.Xe-0X (posix)
+		      $line1 =~ s/(e[-+])0([0-9][0-9])/$1$2/g; #hack on windows X.Xe-00X -> X.Xe-0X (posix)
                       $nerror++ unless ("$line2" eq "$line1");
                     }
                     while (my $line2=<WARN2>) {#FILE2 has more lines
@@ -382,7 +382,7 @@ sub test_systems {
 			push(@error_detail, "checkvir.out ($nerr_vir errors)") if ($nerr_vir > 0);
 
 			$nerror |= $nerr_pot | $nerr_vir;
-			unlink($potout,$poterr);
+			unlink($potout,$poterr) unless $nerr_pot;
 		    }
 		    my $reftrr = "${ref}.trr";
 		    if (! -f $reftrr ) {
